-Deleted busines_settings.py -Fixed example for yookassa config
