+++ backend/payments/config/settings/business_settings.py
@@ -1,6 +1,9 @@
 from datetime import timedelta
 from decimal import Decimal
 
+from pydantic import BaseSettings
+from yookassa import Configuration
+
 MAX_BALANCE_DIGITS = 11
 DEFAULT_CURRENCY = 'RUB'
 SUPPORTED_CURRENCIES = (DEFAULT_CURRENCY,)
@@ -10,5 +13,16 @@ PERIOD_FOR_MYSELF_TASK = timedelta(days=1)
 PERIOD_FOR_GIFT_TASK = timedelta(days=7)
 MINIMUM_PAYOUT_AMOUNT = Decimal(500)
 MAXIMUM_YOOKASSA_PAYOUT = Decimal(500000)
-MAXIMUM_PAYOUT_AMOUNTS = 1
 MAXIMUM_PAYOUTS_PER_MONTH = 1
+
+
+class YookassaConfig(BaseSettings):
+    SHOP_ACCOUNT_ID: int
+    SHOP_SECRET_KEY: str
+
+    def get_payment_settings(self):
+        Configuration.account_id = self.SHOP_ACCOUNT_ID
+        Configuration.secret_key = self.SHOP_SECRET_KEY
+
+
+YOOKASSA_CONFIG = YookassaConfig()
